+++ src/components/screens/chat/chatFormBlocks/ChatFormMain.js
@@ -1,0 +1,26 @@
+import MessagesList from "../MessagesList";
+import { getActiveConversationMessages } from "../../../../store/Messages";
+import { useSelector } from "react-redux";
+
+import { ReactComponent as EmptyChat } from "./../../../../assets/icons/chatForm/EmptyChat.svg";
+
+export default function ChatFormMain({ scrollRef, open }) {
+  const messages = useSelector(getActiveConversationMessages);
+
+  console.log("render --- Main");
+
+  return (
+    <div className="chat-form-main">
+      {!messages.length ? (
+        <div className="chat-empty">
+          <EmptyChat />
+          <p>Please type your message...</p>
+        </div>
+      ) : (
+        <div id="chatMessagesScrollable">
+          <MessagesList scrollRef={scrollRef} openModalFunc={open} />
+        </div>
+      )}
+    </div>
+  );
+}
