+++ src/styles/pages/ChatForm.css
@@ -321,10 +321,6 @@
 }
 
 @media (max-width: 800px) {
-  #chatMessagesScrollable {
-    position: relative;
-    z-index: 2000;
-  }
   .chat-recipient-photo {
     width: 50px;
     height: 50px;
